+++ packages/core/solidity/src/utils/imports-libraries.ts
@@ -1,0 +1,5 @@
+import type { Contract } from '../contract';
+
+export function importsLibrary(contract: Contract, library: string) {
+  return contract.imports.some(i => i.path.startsWith(library));
+}