+++ packages/ui/src/main.ts
@@ -72,7 +72,7 @@ function evaluateSelection(
         return { compatible: true, appType: 'cairo' };
       } else if (
         requestedVersion === 'alpha' ||
-        (semver.satisfies(requestedVersion, cairoAlphaSemver) && cairoVersion !== cairoAlphaVersion)
+        (semver.satisfies(requestedVersion, cairoAlphaSemver) && semver.compare(cairoVersion, cairoAlphaVersion) !== 0)
       ) {
         return { compatible: true, appType: 'cairo_alpha' };
       } else if (requestedVersion === 'stable' || semver.satisfies(requestedVersion, cairoSemver)) {